Move the concept of the track menu actions to a session
